// src/models/book.ts
export interface Book {
    bookId: number;
    authors: string[];
    datetime: string;
    contents: string;
    isbn: string;
    price: number;
    publisher: string;
    sale_price: number;
    status: string;
    thumbnail: string;
    title: string;
    translators: string[];
    url: string;
  }
  